feat(dashboard): redesign verify-OTP to match login (shared AuthShell)
Extracts the split-screen branded panel into a reusable <AuthShell> so /login
and /verify-otp are one continuous experience — identical cinematic left panel
(constellation + glow + logo + feature reveal), swapped right-hand form.

- AuthShell: the shared left panel + right slot + AuthLoader. login/page.tsx now
  just renders <AuthShell><LoginPanel/></AuthShell> (117 lines lighter).
- OtpPanel: the 2FA form on the right — 6-digit input with paste, arrow-key nav,
  auto-submit on completion, resend with cooldown, and session-expiry recovery.
  verify-otp/page.tsx drops to a thin session-gate + <AuthShell><OtpPanel/></AuthShell>.

tsc --noEmit 0 errors; /login and /verify-otp both compile + serve 200 in dev.
